유저 이미지 삭제 구현

+++ back/src/controllers/userAuthController.js
@@ -1,6 +1,5 @@
 const express = require("express");
 
-const { loginRequired } = require("../middlewares/login_required");
 const { userAuthService } = require("../services/userAuthService");
 
 const userAuthController = {
@@ -8,17 +7,17 @@ const userAuthController = {
   postAddUser: async (req, res, next) => {
     try {
       const { email, password, name } = req.body;
-  
+
       // 서비스 파일에서 addUser 함수 실행
       const userInfo = await userAuthService.addUserInfo({
         email,
         password,
         name,
       });
-  
+
       // 서비스에서 에러가 있다면 에러 통보
       if (userInfo.errorMessage) throw new Error("회원가입 실패");
-  
+
       // userInfo를 promise로 반환 하여 전달
       res.status(201).json(userInfo);
     } catch (err) {
@@ -29,13 +28,13 @@ const userAuthController = {
   postLogin: async (req, res, next) => {
     try {
       const { email, password } = req.body;
-  
+
       // 서비스 파일에서 login 함수 실행
       const userLoginInfo = await userAuthService.login(email, password);
-  
+
       // 서비스에서 에러가 있다면 에러 통보
       if (userLoginInfo.errorMessage) throw new Error("로그인 실패");
-  
+
       res.status(201).send(userLoginInfo);
     } catch (err) {
       next(err);
@@ -45,14 +44,14 @@ const userAuthController = {
   getUser: async (req, res, next) => {
     try {
       const { userId } = req.params;
-  
+
       // 서비스 파일에서 getUserInfo 함수 실행
       const currentUserInfo = await userAuthService.getUserInfo(userId);
-  
+
       // 서비스에서 에러가 있다면 에러 통보
       if (currentUserInfo.errorMessage)
         throw new Error("회원 정보 불러오기 실패");
-  
+
       res.status(200).send(currentUserInfo);
     } catch (err) {
       next(err);
@@ -74,13 +73,16 @@ const userAuthController = {
 
       // 변경할 정보를 toUpdate에 초기화
       const toUpdate = { password, imageUrl };
-  
+
       // 서비스 파일에서 updateUser 함수 실행
-      const updatedUser = await userAuthService.updateUserInfo({ userId, toUpdate });
-  
+      const updatedUser = await userAuthService.updateUserInfo({
+        userId,
+        toUpdate,
+      });
+
       // 서비스에서 에러가 있다면 에러 통보
       if (updatedUser.errorMessage) throw new Error("회원 정보 불러오기 실패");
-  
+
       res.status(200).json(updatedUser);
     } catch (err) {
       next(err);
@@ -92,21 +94,20 @@ const userAuthController = {
       const { userId } = req.params;
       const deletedUser = await userAuthService.deleteUserInfo(userId);
 
-      console.log(`컨트롤 확인`, !deletedUser.errorMessage)
+      console.log(`컨트롤 확인`, !deletedUser.errorMessage);
 
-      if (!deletedUser.errorMessage) {
-        console.log(`컨트롤 deletedUser.imageUrl:`,deletedUser.imageUrl);
+      if (!deletedUser.errorMessage && deletedUser.imageUrl) {
+        console.log(`컨트롤 deletedUser.imageUrl:`, deletedUser.imageUrl);
+
         await userAuthService.deleteUserImage(deletedUser.imageUrl);
       }
       if (deletedUser.errorMessage) throw new Error("회원 삭제 실패");
-  
+
       res.status(200).json(deletedUser);
     } catch (err) {
       next(err);
     }
   },
 };
-
-
 
 exports.userAuthController = userAuthController;
